config: Fix compiler datatype comparison warning
/home/laura/git/zproxy/src/config.cpp: In function ‘int parseRedirect(zproxy_backend_redirect*, char*, regmatch_t*, bool)’:
/home/laura/git/zproxy/src/config.cpp:562:11: warning: comparison of integer expressions of different signedness: ‘int’ and ‘const size_t’ {aka ‘const long unsigned int’} [-Wsign-compare]
       i+1 < lin_size && // to avoid a segfault in the next condition
       ~~~~^~~~~~~~~~

+++ src/config.cpp
@@ -557,7 +557,7 @@ static int parseRedirect(struct zproxy_backend_redirect * be, char *lin,
 	be->dynamic = false;
 
 	const size_t lin_size = strlen(lin);
-	for (int i = 0; lin[i] != '\0'; ++i) {
+	for (size_t i = 0; lin[i] != '\0'; ++i) {
 		if (lin[i] == '$' &&
 		    i+1 < lin_size && // to avoid a segfault in the next condition
 		    IN_RANGE(lin[i+1], '1', '9')) {
